Move some Python interface functions from AI to common wrappers
