+++ python/CMakeLists.txt
@@ -8,6 +8,7 @@ target_sources(freeoriond
         ${CMAKE_CURRENT_LIST_DIR}/SetWrapper.h
     PRIVATE
         ${CMAKE_CURRENT_LIST_DIR}/CommonFramework.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/ConfigWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/EmpireWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/EnumWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LoggingWrapper.cpp
@@ -21,6 +22,7 @@ target_sources(freeorionca
         ${CMAKE_CURRENT_LIST_DIR}/SetWrapper.h
     PRIVATE
         ${CMAKE_CURRENT_LIST_DIR}/CommonFramework.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/ConfigWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/EmpireWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/EnumWrapper.cpp
         ${CMAKE_CURRENT_LIST_DIR}/LoggingWrapper.cpp
